+++ masked4j-core/src/main/java/io/github/masked4j/exception/MaskingFailureStrategy.java
@@ -1,0 +1,7 @@
+package io.github.masked4j.exception;
+
+public enum MaskingFailureStrategy {
+  FAIL_FAST,
+  IGNORE,
+  REPLACE_WITH_NULL
+}