+++ masked4j-core/src/main/java/io/github/masked4j/exception/MaskingProcessException.java
@@ -1,0 +1,11 @@
+package io.github.masked4j.exception;
+
+public class MaskingProcessException extends MaskingException {
+  public MaskingProcessException(String message) {
+    super(message);
+  }
+
+  public MaskingProcessException(String message, Throwable cause) {
+    super(message, cause);
+  }
+}